vec3: changed to store floats instead of doubles. tests pass

+++ saralib/vec3.h
@@ -6,18 +6,18 @@
 
 class vec3{
     public:
-        double e[3];
+        float e[3];
 
         vec3() : e{0,0,0} {}
-        vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
+        vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}
 
-        double x() const {return e[0];}
-        double y() const {return e[1];}
-        double z() const {return e[2];}
+        float x() const {return e[0];}
+        float y() const {return e[1];}
+        float z() const {return e[2];}
 
         vec3 operator-() const {return vec3(-e[0], -e[1], -e[2]);}
-        double operator[](int i) const {return e[i];}
-        double& operator[](int i){return e[i];}
+        float operator[](int i) const {return e[i];}
+        float& operator[](int i){return e[i];}
 
         vec3& operator+=(const vec3& v){
             e[0] += v.e[0];
@@ -30,11 +30,11 @@ class vec3{
             return *this *= 1/t;
         }
 
-        double length() const {
+        float length() const {
             return std::sqrt(length_squared());
         }
 
-        double length_squared() const{
+        float length_squared() const{
             return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
         }
 };
@@ -72,7 +72,7 @@ inline vec3 operator/(const vec3&v, double t){
     return (1/t) * v;
 }
 
-inline double dot(const vec3&u, const vec3&v){
+inline float dot(const vec3&u, const vec3&v){
     return u.e[0] * v.e[0] + u.e[1]*v.e[1] + u.e[2]*v.e[2];
 }
